+++ packages/asc-ui/src/components/Checkbox/CheckboxStyle.ts
@@ -1,6 +1,5 @@
 import styled, { css } from 'styled-components'
 import { svgFill, themeColor } from '../../utils'
-import { focusOutline, outlineWhenFocused } from '../../utils/focus'
 import { IconStyle } from '../Icon'
 
 type CheckboxVariant = 'primary' | 'secondary' | 'tertiary'
@@ -48,15 +47,14 @@ const getVariant = () => ({ variant }: Props) => {
 export default styled.input.attrs({
   type: 'checkbox',
 })<Props>`
-  top: 0;
-  left: 0;
-  width: 100%;
   cursor: inherit;
-  height: 100%;
   margin: 0;
-  opacity: 0;
   padding: 0;
   position: absolute;
+  width: 24px;
+  height: 24px;
+  left: 6px;
+  top: 6px;
 `
 
 const CheckboxIconStyle = styled(IconStyle)<Props>`
@@ -68,6 +66,7 @@ const CheckboxIconStyle = styled(IconStyle)<Props>`
   position: relative;
   justify-content: center;
   align-items: center;
+  z-index: 1;
   background-color: ${themeColor('tint', 'level1')};
   transition: background-color 0.2s ease-in-out;
   ${({ checked, indeterminate }) =>
@@ -77,7 +76,7 @@ const CheckboxIconStyle = styled(IconStyle)<Props>`
     `};
 `
 
-const CheckboxWrapperStyle = styled.div<Props & { focus: boolean }>`
+const CheckboxWrapperStyle = styled.div<Props>`
   position: relative;
   display: inline-flex;
   user-select: none;
@@ -86,22 +85,14 @@ const CheckboxWrapperStyle = styled.div<Props & { focus: boolean }>`
   padding: 6px;
   flex-shrink: 0; /* IE11 fix */
   color: ${themeColor('tint', 'level5')};
-  ${({ focus }) =>
-    focus &&
-    css`
-      ${CheckboxIconStyle} {
-        ${focusOutline()}
-      }
-    `}
   ${({ disabled }) =>
     disabled &&
     css`
       pointer-events: none;
     `}
-  ${({ checked, disabled, focus }) =>
+  ${({ checked, disabled }) =>
     !checked &&
     !disabled && // IE11 fix. If an element is disabled, you can still hover on it on IE11.
-    !focus && // don't override the outline if element is also focussed
     css`
       &:hover {
         color: ${themeColor('tint', 'level7')};
@@ -112,18 +103,16 @@ const CheckboxWrapperStyle = styled.div<Props & { focus: boolean }>`
         }
       }
     `}
-    ${({ error, checked, disabled, focus }) =>
+    ${({ error, checked, disabled }) =>
       error &&
       !checked &&
       !disabled &&
-      !focus &&
       css`
         ${CheckboxIconStyle} {
           border-color: red;
           outline: 1px solid red;
         }
       `}
-  ${outlineWhenFocused()}
 `
 
 export { CheckboxWrapperStyle, CheckboxIconStyle }