+++ packages/asc-ui/src/components/Input/InputStyle.ts
@@ -1,6 +1,5 @@
 import styled, { css } from 'styled-components'
 import { themeColor, themeSpacing } from '../../utils'
-import { outlineWhenFocused } from '../../utils/focus'
 import { INPUT_SIZE } from '../shared/constants'
 
 type StyleProps = {
@@ -17,7 +16,6 @@ export const inputCss = css<StyleProps>`
   padding: ${themeSpacing(2)};
   width: 100%;
   height: ${INPUT_SIZE}px;
-  ${outlineWhenFocused()}
   ${({ error }) =>
     !error &&
     css`